fix logger test
