+++ test/CMakeLists.txt
@@ -7,15 +7,15 @@ project(red_engine_test
 
 enable_testing()
 
-set(SOURCES
-        src/main.cpp
-        src/world_test.cpp
-        src/system_test.cpp
-        src/logger_test.cpp)
+# Source files
+file(GLOB_RECURSE SOURCES
+        src/*.cpp
+        )
 
-set(HEADERS
-        include/system_test.hpp
-        include/logger_test.hpp )
+# Headers files
+file(GLOB_RECURSE HEADERS
+        include/*.hpp
+        )
 
 add_executable(red_engine_test ${SOURCES} ${HEADERS})
