+++ test/include/LoggerTest.hpp
