+++ test/src/LoggerTest.cpp
@@ -1,18 +1,18 @@
+#include <LoggerTest.hpp>
 #include <catch2/catch.hpp>
-#include <logger_test.hpp>
 
-TEST_CASE("Logger", "[Logger]")
+TEST_CASE("Logger", "[Debug]")
 {
     LoggerTest logger;
 
     SECTION("Logging trace")
     {
         logger.SetLogLevel(red::LogLevel::LEVEL_TRACE);
-        logger.LogInternal(red::LogLevel::LEVEL_TRACE, 10, "", "Doing stuff with {}, {d} times",
+        logger.LogInternal(red::LogLevel::LEVEL_TRACE, 10, "LoggerTest.cpp", "Doing stuff with {}, {} times",
                            "my bicycle", 10);
         std::cout << logger.lastLog << std::endl;
 
         REQUIRE(logger.lastLog.find("Doing stuff with my bicycle, 10 times") != std::string::npos);
-        REQUIRE(logger.lastLog.find("LEVEL_TRACE") != std::string::npos);
+        REQUIRE(logger.lastLog.find("TRACE") != std::string::npos);
     }
 }
